requirements/all.txt
bleach==1.4.2
blinker==1.4
dominate==2.1.16
Flask-HTTPAuth==2.7.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-PageDown==0.2.1
Flask-Script==2.0.5
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.11
Mako==1.0.3
Markdown==2.6.5
MarkupSafe==0.23
python-editor==0.5
six==1.10.0
visitor==0.1.2
Werkzeug==0.11.2
WTForms==2.0.2
-e git+https://github.com/stepsame/sqlalchemy-continuum.git#egg=sqlalchemy-continuum-dev
